+++ src/main/java/roomescape/global/config/WebMvcConfig.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.global.auth.AuthorizationHeaderInterceptor;
 import roomescape.global.exception.support.BusinessExceptionMappingJackson2HttpMessageConverter;
+import roomescape.reservation.auth.AuthorizationHeaderInterceptor;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
