+++ src/main/java/roomescape/global/exception/NotFoundException.java
@@ -1,6 +1,6 @@
 package roomescape.global.exception;
 
-public abstract class NotFoundException extends BusinessException{
+public abstract class NotFoundException extends BusinessException {
 
     public NotFoundException(String message) {
         super(message);
